+++ tests/testthat/test_batch.R
@@ -3,7 +3,6 @@ source("setup.R")
 context("Batch API Mocking")
 
 
-
 context("v3 API")
 
 
@@ -11,6 +10,7 @@ context("v3 API")
   test_that("v3 multi account batching without flag", {
 
     skip_on_cran()
+    skip_on_travis()
     multi <- google_analytics_3(c(ga_id, ga_id2),
                               start = "2015-07-31", end = "2015-10-01",
                               dimensions=c('medium'),
@@ -26,6 +26,7 @@ context("v3 API")
 
   test_that("v3 multi account batching with flag", {
     skip_on_cran()
+    skip_on_travis()
 
     multi <- google_analytics_3(c(ga_id2, ga_id),
                               start = "2015-07-30", end = "2015-10-01",
@@ -43,6 +44,7 @@ context("v3 API")
   
   test_that("v3 WALK data", {
     skip_on_cran()
+    skip_on_travis()
 
     walked <- suppressWarnings(
       google_analytics_3(ga_id,
@@ -59,6 +61,7 @@ context("v3 API")
 
   test_that("v3 Batch data", {
     skip_on_cran()
+    skip_on_travis()
 
     bb <-       google_analytics_3(ga_id,
                                  start = "2015-07-30", end = "2016-10-01",
@@ -71,20 +74,21 @@ context("v3 API")
 
   })
 
-  # test_that("v4 API matches v3 equivalent API call", {
-  #   skip_on_cran()
-  #   v3 <- google_analytics_v3(ga_id,
-  #                          start = "2015-07-30", end = "2015-10-01",
-  #                          dimensions=c('medium'),
-  #                          metrics = c('sessions'),
-  #                          sort = "ga:sessions")
-  #   v4 <-  google_analytics(ga_id,
-  #                             date_range = c("2015-07-30","2015-10-01"),
-  #                             dimensions=c('medium'),
-  #                             metrics = c('sessions'),
-  #                             order = order_type("sessions"))
-  # 
-  #   expect_equal(v3$medium, v4$medium)
-  #   expect_equal(v3$sessions, v4$sessions)
-  # 
-  # })
+  test_that("v4 API matches v3 equivalent API call", {
+    skip_on_cran()
+    skip_on_travis()
+    v3 <- google_analytics_3(ga_id,
+                           start = "2015-07-30", end = "2015-10-01",
+                           dimensions=c('medium'),
+                           metrics = c('sessions'),
+                           sort = "ga:sessions")
+    v4 <-  google_analytics(ga_id,
+                              date_range = c("2015-07-30","2015-10-01"),
+                              dimensions=c('medium'),
+                              metrics = c('sessions'),
+                              order = order_type("sessions"))
+
+    expect_equal(v3$medium, v4$medium)
+    expect_equal(v3$sessions, v4$sessions)
+
+  })