+++ tests/testthat/test_google_analytics_bq.R
@@ -1,7 +1,7 @@
 context("BigQuery")
 
 test_that("Can make SQL correctly", {
-  
+ 
   expected <- "SELECT trafficSource.source as source, SUM(totals.visits) as sessions FROM (TABLE_DATE_RANGE([foo.ga_sessions_], TIMESTAMP('2017-01-01'), TIMESTAMP('2017-02-01'))) GROUP BY source  LIMIT 100"
   
   got <- google_analytics_bq("blah","foo", 
@@ -14,7 +14,7 @@ test_that("Can make SQL correctly", {
 })
 
 test_that("Error if invalid dimension", {
-  
+
   expect_error(google_analytics_bq("blah","foo", 
                                    start = "2017-01-01", end = "2017-02-01", 
                                    metrics = "sessions", dimensions = "adWordsQuery",