+++ tests/testthat/test_unsampled_report_downloads.R
@@ -9,6 +9,7 @@ test_that("Scope set correctly for Google Drive tests", {
 
 test_that("Can list and download unsampled files", {
   skip_on_cran()
+  skip_on_travis()
   dl <- ga_unsampled_list(accountId = accountId,
                           webPropertyId = webPropId,
                           profileId = ga_id)